ai-service/requirements.txt
fastapi==0.115.6
uvicorn==0.30.6
pymupdf==1.24.9
python-pptx==0.6.23
requests==2.32.3
sentence-transformers==2.2.2
numpy==1.24.3
